Fix ratings handling
git-svn-id: https://svn.apache.org/repos/asf/sling/trunk@1755802 13f79535-47bb-0310-9956-ffa450edef68

+++ samples/slingshot/src/main/java/org/apache/sling/sample/slingshot/ratings/impl/RatingPostServlet.java
@@ -75,7 +75,7 @@ public class RatingPostServlet extends SlingAllMethodsServlet {
 
             final Resource reqResource = resolver.getResource(request.getResource().getPath());
 
-            ratingsService.setRating(reqResource, userId, Integer.valueOf(rating));
+            ratingsService.setRating(reqResource, userId, Double.valueOf(rating));
 
         } catch ( final LoginException le ) {
             throw new ServletException("Unable to login", le);
